refactor: corrections after second code review
