+++ src/data/detail-page/descriptions/devOps.ts
@@ -1,4 +1,4 @@
 export const devOps = `A DevOps engineers combine knowledge about application coding, 
-  'maintenance and management. They are responsible for deployment, monitoring and scaling 
-  'duties necessary during releasing process. They also have in-depth knowledge about automatization, 
-  'operating systems and strong programming skills.`;
+maintenance and management. They are responsible for deployment, monitoring and scaling 
+duties necessary during releasing process. They also have in-depth knowledge about automatization, 
+operating systems and strong programming skills.`;
