+++ src/data/detail-page/resources/ResourceTypeNames.ts
@@ -1,0 +1,12 @@
+export enum ResourceTypeNames {
+  books = 'Books',
+  courses = 'Online courses',
+  platforms = 'Platforms',
+  tools = 'Useful tools',
+  communities = 'Communities',
+  portalsAndPodcasts = 'Portals and Podcasts',
+  games = 'Games & Coding Challenges',
+  varia = 'Varia'
+}
+
+export type ResourceType = keyof typeof ResourceTypeNames;
