--- src/data/detail-page/resources/ResourcesTypeNames.ts
@@ -1,12 +1,0 @@
-import { ResourcesType } from './ResourcesInterfaces';
-
-export const ResourcesTypeNames: Record<ResourcesType, string> = {
-  books: 'Books',
-  courses: 'Online courses',
-  platforms: 'Platforms',
-  tools: 'Useful tools',
-  communities: 'Communities',
-  portalsAndPodcasts: 'Portals and Podcasts',
-  games: 'Games & Coding Challenges',
-  varia: 'Varia',
-};
